Clientクラスに移植すべき機能の移植 一部スペルの修正
