+++ src/Main.java
@@ -2,7 +2,7 @@
 // then press Enter. You can now see whitespace characters in your code.
 public class Main {
     public static void main(String args[]) {
-        Game frame = new Game("MyTitle");
+        Client frame = new Client("MyTitle");
         frame.setVisible(true);
